Any widget now is draggable
The Widget Window has been updated to use the base implementation.
It is allowed to specify the DragHandle that will launch the drag.
Blocking drag per direction is allowed.
